1. Deploy the project to IntelliJ. Use Maven to install required external libraries
2. Use Apache DBCP as the new Java Connection Pool module.

3. Now the main script will provide a connection pool and a connection object. The object will be referred to widgets.

4. The script now work with Mysql instance (in docker) in GCP VM.

This might not be a bug-free commit, need more time to verify.

+++ src/com/ibm/stg/pc/sym/si/util/PainItem.java
@@ -44,18 +44,18 @@ public class PainItem{
 			this.changepercentage  = (float)(this.current - this.baseline)/this.baseline;
 			//System.out.println("pmatrix LHS = "+pmatrix.get(0)[t] + " RHS = "+pmatrix.get(0)[t+1]);
 			if (changepercentage <= pmatrix.get(0)[0] ){
-				this.score =  (float)pmatrix.get(index)[0];
+				this.score = pmatrix.get(index)[0];
 				return score;
 			}
 			
 			if (changepercentage > pmatrix.get(0)[pmatrix.get(index).length -1] ){
-				this.score =  (float)pmatrix.get(index)[pmatrix.get(index).length -1];
+				this.score = pmatrix.get(index)[pmatrix.get(index).length -1];
 				return score;
 			}
 			
 			if( (pmatrix.get(0)[t] <= changepercentage) && (changepercentage <= pmatrix.get(0)[t+1])){
 				//System.out.println("hit t = "+t+" index = "+index);
-				this.score =  (float)pmatrix.get(index)[t+1];
+				this.score = pmatrix.get(index)[t+1];
 				return score;
 			}
 		}
